#731 Dataset playback updated to work with multiple patterns
I think that finally dataset_playback is done (for now...)

As I said, I fully reconfigured the `graphics` dictionary.
It now has the keys:

- colormap - it defines the colormaps for each pattern for rgb images and for each collection
- sensors - it defines the publishers and topics to publish for each sensor
- ros - same as before
- labels - has the labels organized by pattern_key

I have some things that, unfortunately, I could not change now, as I believe we needed to be more developed in the multi-pattern approach.

Some things are still to be done and are marked as TODO on comments:
- Depth and lidar manual labeling only works for a single pattern, as we would need an external input to define which pattern we were labeling
- The infrastructure for defining different colormaps for each pattern is there, but it is not yet implemented