#706 T-SNE analysis addition
Addition of a T-SNE analysis.
Generates the TSNE embedding for the joint states of each collection.
This then is plotted and colored accordingly to the residual of that collection at the end of the calibration.
The objective is to figure out if there is any connection between the joint states and the errors.