#940 Reformating execute and adding it to batch execution
The function execute from atom_core.system has been reformulated.
It now prints in "real time" what the subprocess is outputting.
It is also now responsible to write stdout and stderr to a file.
This has been done in a way that saving in independent of blocking and the filename can have extra addons.

Then, this has been added to configure and to batch execution.
The configure became significantly simpler, as most local functions have been delegated to this one.
Batch execution suffered some changed to adapt to this, but it is now fully functional.
To achieve this, the function removeColorsFromText needed to be moved from atom_core.utilities to atom_core.system, to remove circular imports.